crono/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crono"
version = "1.0.0"
description = "A visual cron job scheduler and manager TUI"
authors = [{name = "Crono Team"}]
requires-python = ">=3.10"
dependencies = [
    "textual>=0.52.0",
    "python-croniter>=2.0.0",
    "pydantic>=2.0.0",
]
readme = "README.md"
license = {text = "MIT"}
keywords = ["cron", "scheduler", "tui", "terminal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
crono = "crono.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["crono*"]